feat: generate checksum for cache blobs
Related to #1158

References:
- https://github.com/weaveworks/tf-controller/issues/1158
